ForumProject/requirements.txt
asgiref==3.8.1
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
dill==0.3.8
Django==5.0.6
django-appconf==1.0.6
django-countries==7.6.1
django-cryptography @ git+https://github.com/saurav-codes/django-cryptography@847cade2cc07af8316d59dc0ded709dee708cbc8
django-extensions==3.2.3
django-filter==24.2
django-phonenumbers==1.0.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
isort==5.13.2
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.0.8
phonenumbers==8.13.37
pillow==10.3.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.1.0
python-decouple==3.8
redis==5.0.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
zope.interface==6.3
